De-duplicate fixtures
