+++ tests/components/schlage/conftest.py
@@ -1,11 +1,13 @@
 """Common fixtures for the Schlage tests."""
 from collections.abc import Generator
-from unittest.mock import AsyncMock, patch
+from unittest.mock import AsyncMock, Mock, create_autospec, patch
 
+from pyschlage.lock import Lock
 import pytest
 
 from homeassistant.components.schlage.const import DOMAIN
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
+from homeassistant.core import HomeAssistant
 
 from tests.common import MockConfigEntry
 
@@ -22,6 +24,23 @@ def mock_config_entry() -> MockConfigEntry:
         },
         unique_id="abc123",
     )
+
+
+@pytest.fixture
+async def mock_added_config_entry(
+    hass: HomeAssistant,
+    mock_config_entry: MockConfigEntry,
+    mock_pyschlage_auth: Mock,
+    mock_schlage: Mock,
+    mock_lock: Mock,
+) -> MockConfigEntry:
+    """Mock ConfigEntry that's been added to HA."""
+    mock_schlage.locks.return_value = [mock_lock]
+    mock_config_entry.add_to_hass(hass)
+    await hass.config_entries.async_setup(mock_config_entry.entry_id)
+    await hass.async_block_till_done()
+    assert DOMAIN in hass.config_entries.async_domains()
+    return mock_config_entry
 
 
 @pytest.fixture
@@ -46,3 +65,19 @@ def mock_pyschlage_auth():
     with patch("pyschlage.Auth", autospec=True) as mock_auth:
         mock_auth.return_value.user_id = "abc123"
         yield mock_auth.return_value
+
+
+@pytest.fixture
+def mock_lock():
+    """Mock Lock fixture."""
+    mock_lock = create_autospec(Lock)
+    mock_lock.configure_mock(
+        device_id="test",
+        name="Vault Door",
+        model_name="<model-name>",
+        is_locked=False,
+        is_jammed=False,
+        battery_level=0,
+        firmware_version="1.0",
+    )
+    return mock_lock
